+++ instrumentation/opentelemetry-instrumentation-asgi/src/opentelemetry/instrumentation/asgi/__init__.py
@@ -201,6 +201,31 @@ from typing import Any, Awaitable, Callable, DefaultDict, Tuple
 from asgiref.compatibility import guarantee_single_callable
 
 from opentelemetry import context, trace
+from opentelemetry.instrumentation._semconv import (
+    _filter_semconv_active_request_count_attr,
+    _filter_semconv_duration_attrs,
+    _get_schema_url,
+    _HTTPStabilityMode,
+    _OpenTelemetrySemanticConventionStability,
+    _OpenTelemetryStabilitySignalType,
+    _report_new,
+    _report_old,
+    _server_active_requests_count_attrs_new,
+    _server_active_requests_count_attrs_old,
+    _server_duration_attrs_new,
+    _server_duration_attrs_old,
+    _set_http_flavor_version,
+    _set_http_host,
+    _set_http_method,
+    _set_http_net_host_port,
+    _set_http_peer_ip,
+    _set_http_peer_port_server,
+    _set_http_scheme,
+    _set_http_target,
+    _set_http_url,
+    _set_http_user_agent,
+    _set_status,
+)
 from opentelemetry.instrumentation.asgi.types import (
     ClientRequestHook,
     ClientResponseHook,
@@ -210,27 +235,31 @@ from opentelemetry.instrumentation.asgi.version import __version__  # noqa
 from opentelemetry.instrumentation.propagators import (
     get_global_response_propagator,
 )
-from opentelemetry.instrumentation.utils import (
-    _start_internal_or_server_span,
-    http_status_to_status_code,
-)
+from opentelemetry.instrumentation.utils import _start_internal_or_server_span
 from opentelemetry.metrics import get_meter
 from opentelemetry.propagators.textmap import Getter, Setter
+from opentelemetry.semconv._incubating.metrics.http_metrics import (
+    create_http_server_active_requests,
+    create_http_server_request_body_size,
+    create_http_server_response_body_size,
+)
 from opentelemetry.semconv.metrics import MetricInstruments
+from opentelemetry.semconv.metrics.http_metrics import (
+    HTTP_SERVER_REQUEST_DURATION,
+)
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.trace import set_span_in_context
-from opentelemetry.trace.status import Status, StatusCode
 from opentelemetry.util.http import (
     OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS,
     OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_REQUEST,
     OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_RESPONSE,
     SanitizeValue,
-    _parse_active_request_count_attrs,
-    _parse_duration_attrs,
+    _parse_url_query,
     get_custom_headers,
     normalise_request_header_name,
     normalise_response_header_name,
     remove_url_credentials,
+    sanitize_method,
 )
 
 
@@ -295,7 +324,10 @@ class ASGISetter(Setter[dict]):
 asgi_setter = ASGISetter()
 
 
-def collect_request_attributes(scope):
+# pylint: disable=too-many-branches
+def collect_request_attributes(
+    scope, sem_conv_opt_in_mode=_HTTPStabilityMode.DEFAULT
+):
     """Collects HTTP request attributes from the ASGI scope and returns a
     dictionary to be used as span creation attributes."""
     server_host, port, http_url = get_host_port_url_tuple(scope)
@@ -304,31 +336,52 @@ def collect_request_attributes(scope):
         if isinstance(query_string, bytes):
             query_string = query_string.decode("utf8")
         http_url += "?" + urllib.parse.unquote(query_string)
+    result = {}
 
-    result = {
-        SpanAttributes.HTTP_SCHEME: scope.get("scheme"),
-        SpanAttributes.HTTP_HOST: server_host,
-        SpanAttributes.NET_HOST_PORT: port,
-        SpanAttributes.HTTP_FLAVOR: scope.get("http_version"),
-        SpanAttributes.HTTP_TARGET: scope.get("path"),
-        SpanAttributes.HTTP_URL: remove_url_credentials(http_url),
-    }
-    http_method = scope.get("method")
+    scheme = scope.get("scheme")
+    if scheme:
+        _set_http_scheme(result, scheme, sem_conv_opt_in_mode)
+    if server_host:
+        _set_http_host(result, server_host, sem_conv_opt_in_mode)
+    if port:
+        _set_http_net_host_port(result, port, sem_conv_opt_in_mode)
+    flavor = scope.get("http_version")
+    if flavor:
+        _set_http_flavor_version(result, flavor, sem_conv_opt_in_mode)
+    path = scope.get("path")
+    if path:
+        _set_http_target(
+            result, path, path, query_string, sem_conv_opt_in_mode
+        )
+    if http_url:
+        _set_http_url(
+            result, remove_url_credentials(http_url), sem_conv_opt_in_mode
+        )
+
+    http_method = scope.get("method", "")
     if http_method:
-        result[SpanAttributes.HTTP_METHOD] = http_method
+        _set_http_method(
+            result,
+            http_method,
+            sanitize_method(http_method),
+            sem_conv_opt_in_mode,
+        )
 
     http_host_value_list = asgi_getter.get(scope, "host")
     if http_host_value_list:
-        result[SpanAttributes.HTTP_SERVER_NAME] = ",".join(
-            http_host_value_list
-        )
+        if _report_old(sem_conv_opt_in_mode):
+            result[SpanAttributes.HTTP_SERVER_NAME] = ",".join(
+                http_host_value_list
+            )
     http_user_agent = asgi_getter.get(scope, "user-agent")
     if http_user_agent:
-        result[SpanAttributes.HTTP_USER_AGENT] = http_user_agent[0]
+        _set_http_user_agent(result, http_user_agent[0], sem_conv_opt_in_mode)
 
     if "client" in scope and scope["client"] is not None:
-        result[SpanAttributes.NET_PEER_IP] = scope.get("client")[0]
-        result[SpanAttributes.NET_PEER_PORT] = scope.get("client")[1]
+        _set_http_peer_ip(result, scope.get("client")[0], sem_conv_opt_in_mode)
+        _set_http_peer_port_server(
+            result, scope.get("client")[1], sem_conv_opt_in_mode
+        )
 
     # remove None values
     result = {k: v for k, v in result.items() if v is not None}
@@ -367,11 +420,18 @@ def get_host_port_url_tuple(scope):
     server = scope.get("server") or ["0.0.0.0", 80]
     port = server[1]
     server_host = server[0] + (":" + str(port) if str(port) != "80" else "")
-    full_path = scope.get("root_path", "") + scope.get("path", "")
+    # using the scope path is enough, see:
+    # - https://asgi.readthedocs.io/en/latest/specs/www.html#http-connection-scope (see: root_path and path)
+    # - https://asgi.readthedocs.io/en/latest/specs/www.html#wsgi-compatibility (see: PATH_INFO)
+    #       PATH_INFO can be derived by stripping root_path from path
+    #       -> that means that the path should contain the root_path already, so prefixing it again is not necessary
+    # - https://wsgi.readthedocs.io/en/latest/definitions.html#envvar-PATH_INFO
+    full_path = scope.get("path", "")
     http_url = scope.get("scheme", "http") + "://" + server_host + full_path
     return server_host, port, http_url
 
 
+<<<<<<< middleware-always-set-status-code-on-duration-attrs
 def maybe_inject_context(message, server_span, setter):
     """Inject the context if there's a global propagator."""
     propagator = get_global_response_propagator()
@@ -386,23 +446,32 @@ def maybe_inject_context(message, server_span, setter):
 
 
 def set_status_code(span, status_code):
+=======
+def set_status_code(
+    span,
+    status_code,
+    metric_attributes=None,
+    sem_conv_opt_in_mode=_HTTPStabilityMode.DEFAULT,
+):
+>>>>>>> main
     """Adds HTTP response attributes to span using the status_code argument."""
     if not span.is_recording():
         return
+    status_code_str = str(status_code)
+
     try:
         status_code = int(status_code)
     except ValueError:
-        span.set_status(
-            Status(
-                StatusCode.ERROR,
-                "Non-integer HTTP status: " + repr(status_code),
-            )
-        )
-    else:
-        span.set_attribute(SpanAttributes.HTTP_STATUS_CODE, status_code)
-        span.set_status(
-            Status(http_status_to_status_code(status_code, server_span=True))
-        )
+        status_code = -1
+    if metric_attributes is None:
+        metric_attributes = {}
+    _set_status(
+        span,
+        metric_attributes,
+        status_code,
+        status_code_str,
+        sem_conv_opt_in_mode,
+    )
 
 
 def get_default_span_details(scope: dict) -> Tuple[str, dict]:
@@ -417,7 +486,9 @@ def get_default_span_details(scope: dict) -> Tuple[str, dict]:
         a tuple of the span name, and any attributes to attach to the span.
     """
     path = scope.get("path", "").strip()
-    method = scope.get("method", "").strip()
+    method = sanitize_method(scope.get("method", "").strip())
+    if method == "_OTHER":
+        method = "HTTP"
     if method and path:  # http
         return f"{method} {path}", {}
     if path:  # websocket
@@ -491,13 +562,18 @@ class OpenTelemetryMiddleware:
         http_capture_headers_server_response: list[str] | None = None,
         http_capture_headers_sanitize_fields: list[str] | None = None,
     ):
+        # initialize semantic conventions opt-in if needed
+        _OpenTelemetrySemanticConventionStability._initialize()
+        sem_conv_opt_in_mode = _OpenTelemetrySemanticConventionStability._get_opentelemetry_stability_opt_in_mode(
+            _OpenTelemetryStabilitySignalType.HTTP,
+        )
         self.app = guarantee_single_callable(app)
         self.tracer = (
             trace.get_tracer(
                 __name__,
                 __version__,
                 tracer_provider,
-                schema_url="https://opentelemetry.io/schemas/1.11.0",
+                schema_url=_get_schema_url(sem_conv_opt_in_mode),
             )
             if tracer is None
             else tracer
@@ -507,30 +583,51 @@ class OpenTelemetryMiddleware:
                 __name__,
                 __version__,
                 meter_provider,
-                schema_url="https://opentelemetry.io/schemas/1.11.0",
+                schema_url=_get_schema_url(sem_conv_opt_in_mode),
             )
             if meter is None
             else meter
         )
-        self.duration_histogram = self.meter.create_histogram(
-            name=MetricInstruments.HTTP_SERVER_DURATION,
-            unit="ms",
-            description="Duration of HTTP client requests.",
-        )
-        self.server_response_size_histogram = self.meter.create_histogram(
-            name=MetricInstruments.HTTP_SERVER_RESPONSE_SIZE,
-            unit="By",
-            description="measures the size of HTTP response messages (compressed).",
-        )
-        self.server_request_size_histogram = self.meter.create_histogram(
-            name=MetricInstruments.HTTP_SERVER_REQUEST_SIZE,
-            unit="By",
-            description="Measures the size of HTTP request messages (compressed).",
-        )
-        self.active_requests_counter = self.meter.create_up_down_counter(
-            name=MetricInstruments.HTTP_SERVER_ACTIVE_REQUESTS,
-            unit="requests",
-            description="measures the number of concurrent HTTP requests that are currently in-flight",
+        self.duration_histogram_old = None
+        if _report_old(sem_conv_opt_in_mode):
+            self.duration_histogram_old = self.meter.create_histogram(
+                name=MetricInstruments.HTTP_SERVER_DURATION,
+                unit="ms",
+                description="Duration of HTTP server requests.",
+            )
+        self.duration_histogram_new = None
+        if _report_new(sem_conv_opt_in_mode):
+            self.duration_histogram_new = self.meter.create_histogram(
+                name=HTTP_SERVER_REQUEST_DURATION,
+                description="Duration of HTTP server requests.",
+                unit="s",
+            )
+        self.server_response_size_histogram = None
+        if _report_old(sem_conv_opt_in_mode):
+            self.server_response_size_histogram = self.meter.create_histogram(
+                name=MetricInstruments.HTTP_SERVER_RESPONSE_SIZE,
+                unit="By",
+                description="measures the size of HTTP response messages (compressed).",
+            )
+        self.server_response_body_size_histogram = None
+        if _report_new(sem_conv_opt_in_mode):
+            self.server_response_body_size_histogram = (
+                create_http_server_response_body_size(self.meter)
+            )
+        self.server_request_size_histogram = None
+        if _report_old(sem_conv_opt_in_mode):
+            self.server_request_size_histogram = self.meter.create_histogram(
+                name=MetricInstruments.HTTP_SERVER_REQUEST_SIZE,
+                unit="By",
+                description="Measures the size of HTTP request messages (compressed).",
+            )
+        self.server_request_body_size_histogram = None
+        if _report_new(sem_conv_opt_in_mode):
+            self.server_request_body_size_histogram = (
+                create_http_server_request_body_size(self.meter)
+            )
+        self.active_requests_counter = create_http_server_active_requests(
+            self.meter
         )
         self.excluded_urls = excluded_urls
         self.default_span_details = (
@@ -540,6 +637,7 @@ class OpenTelemetryMiddleware:
         self.client_request_hook = client_request_hook
         self.client_response_hook = client_response_hook
         self.content_length_header = None
+        self._sem_conv_opt_in_mode = sem_conv_opt_in_mode
 
         # Environment variables as constructor parameters
         self.http_capture_headers_server_request = (
@@ -570,6 +668,7 @@ class OpenTelemetryMiddleware:
             or []
         )
 
+    # pylint: disable=too-many-statements
     async def __call__(
         self,
         scope: dict[str, Any],
@@ -593,7 +692,9 @@ class OpenTelemetryMiddleware:
 
         span_name, additional_attributes = self.default_span_details(scope)
 
-        attributes = collect_request_attributes(scope)
+        attributes = collect_request_attributes(
+            scope, self._sem_conv_opt_in_mode
+        )
         attributes.update(additional_attributes)
         span, token = _start_internal_or_server_span(
             tracer=self.tracer,
@@ -604,9 +705,9 @@ class OpenTelemetryMiddleware:
             attributes=attributes,
         )
         active_requests_count_attrs = _parse_active_request_count_attrs(
-            attributes
+            attributes,
+            self._sem_conv_opt_in_mode,
         )
-        duration_attrs = _parse_duration_attrs(attributes)
 
         if scope["type"] == "http":
             self.active_requests_counter.add(1, active_requests_count_attrs)
@@ -642,7 +743,7 @@ class OpenTelemetryMiddleware:
                     span_name,
                     scope,
                     send,
-                    duration_attrs,
+                    attributes,
                 )
 
                 await self.app(scope, otel_receive, otel_send)
@@ -650,16 +751,44 @@ class OpenTelemetryMiddleware:
             if scope["type"] == "http":
                 target = _collect_target_attribute(scope)
                 if target:
-                    duration_attrs[SpanAttributes.HTTP_TARGET] = target
-                duration = max(round((default_timer() - start) * 1000), 0)
-                self.duration_histogram.record(duration, duration_attrs)
+                    path, query = _parse_url_query(target)
+                    _set_http_target(
+                        attributes,
+                        target,
+                        path,
+                        query,
+                        self._sem_conv_opt_in_mode,
+                    )
+                duration_s = default_timer() - start
+                duration_attrs_old = _parse_duration_attrs(
+                    attributes, _HTTPStabilityMode.DEFAULT
+                )
+                if target:
+                    duration_attrs_old[SpanAttributes.HTTP_TARGET] = target
+                duration_attrs_new = _parse_duration_attrs(
+                    attributes, _HTTPStabilityMode.HTTP
+                )
+                if self.duration_histogram_old:
+                    self.duration_histogram_old.record(
+                        max(round(duration_s * 1000), 0), duration_attrs_old
+                    )
+                if self.duration_histogram_new:
+                    self.duration_histogram_new.record(
+                        max(duration_s, 0), duration_attrs_new
+                    )
                 self.active_requests_counter.add(
                     -1, active_requests_count_attrs
                 )
                 if self.content_length_header:
-                    self.server_response_size_histogram.record(
-                        self.content_length_header, duration_attrs
-                    )
+                    if self.server_response_size_histogram:
+                        self.server_response_size_histogram.record(
+                            self.content_length_header, duration_attrs_old
+                        )
+                    if self.server_response_body_size_histogram:
+                        self.server_response_body_size_histogram.record(
+                            self.content_length_header, duration_attrs_new
+                        )
+
                 request_size = asgi_getter.get(scope, "content-length")
                 if request_size:
                     try:
@@ -667,9 +796,14 @@ class OpenTelemetryMiddleware:
                     except ValueError:
                         pass
                     else:
-                        self.server_request_size_histogram.record(
-                            request_size_amount, duration_attrs
-                        )
+                        if self.server_request_size_histogram:
+                            self.server_request_size_histogram.record(
+                                request_size_amount, duration_attrs_old
+                            )
+                        if self.server_request_body_size_histogram:
+                            self.server_request_body_size_histogram.record(
+                                request_size_amount, duration_attrs_new
+                            )
             if token:
                 context.detach(token)
             if span.is_recording():
@@ -688,7 +822,12 @@ class OpenTelemetryMiddleware:
                     self.client_request_hook(receive_span, scope, message)
                 if receive_span.is_recording():
                     if message["type"] == "websocket.receive":
-                        set_status_code(receive_span, 200)
+                        set_status_code(
+                            receive_span,
+                            200,
+                            None,
+                            self._sem_conv_opt_in_mode,
+                        )
                     receive_span.set_attribute(
                         "asgi.event.type", message["type"]
                     )
@@ -697,7 +836,12 @@ class OpenTelemetryMiddleware:
         return otel_receive
 
     def _get_otel_send(
-        self, server_span, server_span_name, scope, send, duration_attrs
+        self,
+        server_span,
+        server_span_name,
+        scope,
+        send,
+        duration_attrs,
     ):
         expecting_trailers = False
 
@@ -722,10 +866,41 @@ class OpenTelemetryMiddleware:
 
                 if send_span.is_recording():
                     if message["type"] == "http.response.start":
+<<<<<<< middleware-always-set-status-code-on-duration-attrs
                         expecting_trailers = message.get("trailers", False)
                     if status_code:
                         set_status_code(server_span, status_code)
                         set_status_code(send_span, status_code)
+=======
+                        status_code = message["status"]
+                        # We record metrics only once
+                        set_status_code(
+                            server_span,
+                            status_code,
+                            duration_attrs,
+                            self._sem_conv_opt_in_mode,
+                        )
+                        set_status_code(
+                            send_span,
+                            status_code,
+                            None,
+                            self._sem_conv_opt_in_mode,
+                        )
+                        expecting_trailers = message.get("trailers", False)
+                    elif message["type"] == "websocket.send":
+                        set_status_code(
+                            server_span,
+                            200,
+                            duration_attrs,
+                            self._sem_conv_opt_in_mode,
+                        )
+                        set_status_code(
+                            send_span,
+                            200,
+                            None,
+                            self._sem_conv_opt_in_mode,
+                        )
+>>>>>>> main
                     send_span.set_attribute("asgi.event.type", message["type"])
                     if (
                         server_span.is_recording()
@@ -774,3 +949,25 @@ class OpenTelemetryMiddleware:
                 server_span.end()
 
         return otel_send
+
+
+def _parse_duration_attrs(
+    req_attrs, sem_conv_opt_in_mode=_HTTPStabilityMode.DEFAULT
+):
+    return _filter_semconv_duration_attrs(
+        req_attrs,
+        _server_duration_attrs_old,
+        _server_duration_attrs_new,
+        sem_conv_opt_in_mode,
+    )
+
+
+def _parse_active_request_count_attrs(
+    req_attrs, sem_conv_opt_in_mode=_HTTPStabilityMode.DEFAULT
+):
+    return _filter_semconv_active_request_count_attr(
+        req_attrs,
+        _server_active_requests_count_attrs_old,
+        _server_active_requests_count_attrs_new,
+        sem_conv_opt_in_mode,
+    )